+++ backend_rust/src/adapters/repository.rs
@@ -1,4 +1,1 @@
-pub mod db_process_registration_repository;
-pub mod db_recipe_repository;
-pub mod db_resource_repository;
 pub mod mysql;
