+++ backend_rust/src/adapters/repository/mysql.rs
@@ -1,3 +1,6 @@
+pub mod db_process_registration_repository;
+pub mod db_recipe_repository;
+pub mod db_resource_repository;
 pub mod entity;
 pub mod recipe_mapper;
 pub mod resource_mapper;
