+++ backend_rust/src/adapters/repository/mysql/db_process_registration_repository.rs
