+++ backend_rust/src/main.rs
@@ -3,9 +3,9 @@ use async_graphql::{EmptySubscription, Schema};
 use axum::routing::post;
 use axum::{Router, extract::Extension};
 use cpp_backend::adapters::grpc::cooking::process_service_client;
-use cpp_backend::adapters::repository::db_process_registration_repository::DbProcessRepository;
-use cpp_backend::adapters::repository::db_recipe_repository::DbRecipeRepository;
-use cpp_backend::adapters::repository::db_resource_repository::DbResourceRepository;
+use cpp_backend::adapters::repository::mysql::db_process_registration_repository::DbProcessRepository;
+use cpp_backend::adapters::repository::mysql::db_recipe_repository::DbRecipeRepository;
+use cpp_backend::adapters::repository::mysql::db_resource_repository::DbResourceRepository;
 use cpp_backend::presentation::{
     controller::graphql_controller::graphql_handler,
     graphql::{mutation::Mutation, query::Query},
