+++ backend_rust/src/presentation/graphql/query.rs
@@ -2,8 +2,8 @@ use async_graphql::{Context, EmptySubscription, ID, Object, Schema};
 use std::sync::Arc;
 
 use crate::adapters::recipe_mapper::{RecipeDetailMapper, RecipeMapper};
-use crate::adapters::repository::db_recipe_repository::DbRecipeRepository;
-use crate::adapters::repository::db_resource_repository::DbResourceRepository;
+use crate::adapters::repository::mysql::db_recipe_repository::DbRecipeRepository;
+use crate::adapters::repository::mysql::db_resource_repository::DbResourceRepository;
 use crate::application::mapper::api_mapper::ApiMapper;
 use crate::application::usecase::get_all_resources_usecase::GetAllResourcesUsecase;
 use crate::application::usecase::get_one_recipe_by_id_usecase::GetOneRecipeByIdUseCase;
